Перекомпановка классов
